// src/app/models/Product.ts
export interface Product{
    id: number;
    marca: string;
    playera: number;
    blusa: number;
    camisa: number;
    pantalon: number;
    falda: number;
    chamarra: number;
    saco: number;
    sueter: number;
    sudadera: number;
    zapatos: number;
    vestido: number;
    bolsa: number;
    monoEnterizo: number;
    image: string;
}